requirements.txt
flask==3.0.2
requests==2.31.0 
boto3==1.36.1
python-dotenv==1.0.1
tqdm==4.66.2
flask-cors==4.0.0